Refactor dashboard (competitor) and nav bar
