+++ src/components/Navbar.module.css
@@ -5,18 +5,18 @@
 }
 .wrap{
     background: rgba(10,10,10,.8);
-    font-size: 3vw;
-    display: relative;
+    font-size: 2.6vw;
+    display: flex;
+    align-items: center;
+    padding: 0 1vw;
     /* font-weight: bold; */
 }
 
 .wrap{
-    font-family: "CMU Serif", serif;
+    font-family: "Monaspace Xenon", monospace;
     color: #FFF;
 }
 
-.right{
-    position: absolute;
-    top: 1vh;
-    right: 1vh;
+.fill{
+    flex: 1 0 0;
 }
